requirements.txt
Flask
torch
torchvision
clip-by-openai
faiss-cpu
transnetv2
Pillow
